Add Business and Data Visualization settings components to manage application settings
